Criação de classes no dominio e readme para as pastas de camadas

+++ app/src/main/java/com/apphamba/hamba/usuario/dominio/Usuario.java
@@ -1,7 +1,7 @@
 package com.apphamba.hamba.usuario.dominio;
 
 public class Usuario {
-    private int idUsuario;
+    private long idUsuario;
     private String nome;
     private String email;
     private String senha;
